requirements.txt
Flask
requests
python-dotenv
functions-framework==3.*